chore: cleanup `TerminalDrawer`
